Add files via upload
Objective:
The main objective of this project is to design and implement a robust data preprocessing system that addresses common challenges such as missing values, outliers, inconsistent formatting, and noise. By performing effective data preprocessing, the project aims to enhance the quality, reliability, and usefulness of the data for machine learning.

Introduction:
The data used for pre processing is Employee dataset, which consists of 148 rows and 6 columns. The columns include Company, Age, Salary, Place, Country and Gender.

A systematic and comprehensive approach has been employed for data cleaning and preprocessing, utilizing the following steps to ensure its quality and suitability for various predictive modeling tasks.

1.Basic Insights: Analyzing the data types, identifying missing values, and obtaining a statistical summary, has helped in developing a comprehensive understanding of the dataset's structure and content.

2.Data cleaning: Duplicate rows have been removed from the dataset, to ensure data integrity. Additionally, outliers have been addressed, as they can significantly impact the performance of predictive models. Missing values in each column are replaced strategically, considering factors such as the data distribution.

3.Data Analysis: Basic data analysis has been performed to gain deeper insights into the relationships between different features such as Age, Gender and Salary.

4.Data encoding: Data encoding techniques has been employed to facilitate the use of categorical data, by converting them into numerical representations.

5.Feature Scaling: Finally, feature scaling has been applied on numerical predictor variables to standardize their scales for improved model performance.